tr1/output: merge s_output.c to output.c
